third_party/WebKit/LayoutTests/accessibility/is-ignored-change-sends-notification-expected.txt

This test ensures that a change to accessibilityIsIgnored fires a children changed notification on the parent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS accessibleElementById('hiddenDivContainer') != null is true
PASS accessibleElementById('hiddenDiv') != null is false
PASS accessibleElementById('invisibleDivContainer') != null is true
PASS accessibleElementById('invisibleDiv') != null is false
PASS accessibleElementById('emptyDiv') != null is true
PASS accessibleElementById('divWithoutRoleContainer') != null is true


Got ChildrenChanged notification on hiddenDivContainer
PASS accessibleElementById('hiddenDivContainer').childrenCount is 1
Got ChildrenChanged notification on invisibleDivContainer
PASS accessibleElementById('invisibleDivContainer').childrenCount is 1
Got ChildrenChanged notification on emptyDiv
PASS accessibleElementById('emptyDiv').childrenCount is 1
Got ChildrenChanged notification on divWithoutRoleContainer
PASS accessibleElementById('divWithoutRoleContainer').childrenCount is 1
All notifications received successfully.
PASS successfullyParsed is true

TEST COMPLETE
